Move all rpc modules to rpc submodule
This follows more closely upstream master, so we might be able to track the changes there
