+++ networking_arista/ml2/drivers/type_arista_vlan.py
@@ -20,8 +20,8 @@ from oslo_config import cfg
 from oslo_log import log
 
 from networking_arista._i18n import _LI
-from networking_arista.ml2 import arista_ml2
 from networking_arista.ml2.drivers import driver_helpers
+from networking_arista.ml2.rpc.arista_eapi import AristaRPCWrapperEapi
 
 LOG = log.getLogger(__name__)
 
@@ -39,7 +39,7 @@ class AristaVlanTypeDriver(type_vlan.VlanTypeDriver):
 
     def __init__(self):
         super(AristaVlanTypeDriver, self).__init__()
-        self.rpc = arista_ml2.AristaRPCWrapperEapi()
+        self.rpc = AristaRPCWrapperEapi()
         self.sync_service = driver_helpers.VlanSyncService(self.rpc)
         self.network_vlan_ranges = dict()
         self.sync_timeout = cfg.CONF.arista_type_driver['sync_interval']
