+++ networking_arista/ml2/mechanism_arista.py
@@ -34,23 +34,16 @@ from oslo_service import loopingcall
 
 from networking_arista._i18n import _, _LI, _LE
 from networking_arista.common import config  # noqa
+from networking_arista.common import constants
 from networking_arista.common import db
 from networking_arista.common import db_lib
 from networking_arista.common import exceptions as arista_exc
 from networking_arista.ml2 import arista_ml2
+from networking_arista.ml2 import rpc as ml2_rpc
 from networking_arista.ml2 import sec_group_callback
 
 LOG = logging.getLogger(__name__)
-
-# Messages
-EOS_UNREACHABLE_MSG = _('Unable to reach EOS')
-UNABLE_TO_DELETE_PORT_MSG = _('Unable to delete port from EOS')
-UNABLE_TO_DELETE_DEVICE_MSG = _('Unable to delete device')
-
-# Constants
-INTERNAL_TENANT_ID = 'INTERNAL-TENANT-ID'
-PORT_BINDING_HOST = 'binding:host_id'
-MECHANISM_DRV_NAME = 'arista'
+cfg.CONF.import_group('ml2_arista', 'networking_arista.common.config')
 
 
 def pretty_log(tag, obj):
@@ -85,16 +78,17 @@ class AristaDriver(driver_api.MechanismDriver):
             self.rpc = rpc
             self.eapi = rpc
         else:
-            self.eapi = arista_ml2.AristaRPCWrapperEapi(self.ndb)
+            self.eapi = rpc.AristaRPCWrapperEapi(self.ndb)
             api_type = confg['api_type'].upper()
             if api_type == 'EAPI':
                 LOG.info("Using EAPI for RPC")
-                self.rpc = arista_ml2.AristaRPCWrapperEapi(self.ndb)
+                self.rpc = ml2_rpc.arista_eapi.AristaRPCWrapperEapi(self.ndb)
             elif api_type == 'JSON':
                 LOG.info("Using JSON for RPC")
-                self.rpc = arista_ml2.AristaRPCWrapperJSON(self.ndb)
+                self.rpc = ml2_rpc.arista_json.AristaRPCWrapperJSON(self.ndb)
             elif api_type == 'NOCVX':
-                self.rpc = arista_ml2.AristaNoCvxWrapperBase(self.ndb)
+                self.rpc = ml2_rpc.arista_nocvx.AristaNoCvxWrapperBase(
+                    self.ndb)
                 self.eapi = self.rpc
             else:
                 msg = "RPC mechanism %s not recognized" % api_type
@@ -131,7 +125,7 @@ class AristaDriver(driver_api.MechanismDriver):
                 return
 
         network_id = network['id']
-        tenant_id = network['tenant_id'] or INTERNAL_TENANT_ID
+        tenant_id = network['tenant_id'] or constants.INTERNAL_TENANT_ID
         plugin_context = context._plugin_context
         db_lib.remember_tenant(plugin_context, tenant_id)
         for segment in segments:
@@ -147,7 +141,7 @@ class AristaDriver(driver_api.MechanismDriver):
         network = context.current
         network_id = network['id']
         network_name = network['name']
-        tenant_id = network['tenant_id'] or INTERNAL_TENANT_ID
+        tenant_id = network['tenant_id'] or constants.INTERNAL_TENANT_ID
         segments = context.network_segments
         shared_net = network['shared']
 
@@ -194,7 +188,8 @@ class AristaDriver(driver_api.MechanismDriver):
                 new_network['shared'] != orig_network['shared']):
             network_id = new_network['id']
             network_name = new_network['name']
-            tenant_id = new_network['tenant_id'] or INTERNAL_TENANT_ID
+            tenant_id = new_network['tenant_id'] or \
+                constants.INTERNAL_TENANT_ID
             shared_net = new_network['shared']
             if db_lib.is_network_provisioned(plugin_context,
                                              tenant_id, network_id):
@@ -218,7 +213,7 @@ class AristaDriver(driver_api.MechanismDriver):
         """Delete the network information from the DB."""
         network = context.current
         network_id = network['id']
-        tenant_id = network['tenant_id'] or INTERNAL_TENANT_ID
+        tenant_id = network['tenant_id'] or constants.INTERNAL_TENANT_ID
         plugin_context = context._plugin_context
         if db_lib.is_network_provisioned(plugin_context, tenant_id,
                                          network_id):
@@ -243,7 +238,7 @@ class AristaDriver(driver_api.MechanismDriver):
                 # If network type is not VLAN, do nothing
                 return
         network_id = network['id']
-        tenant_id = network['tenant_id'] or INTERNAL_TENANT_ID
+        tenant_id = network['tenant_id'] or constants.INTERNAL_TENANT_ID
 
         # Succeed deleting network in case EOS is not accessible.
         # EOS state will be updated by sync thread once EOS gets
@@ -256,20 +251,6 @@ class AristaDriver(driver_api.MechanismDriver):
             LOG.error(_LE('delete_network_postcommit: Did not delete '
                           'network %(network_id)s. Reason: %(err)s'),
                       {'network_id': network_id, 'err': err})
-
-    def create_port_precommit(self, context):
-        """Remember the information about a VM and its ports
-
-        A VM information, along with the physical host information
-        is saved.
-        """
-
-        # Returning from here, since the update_port_precommit is performing
-        # same operation, and also need of port binding information to decide
-        # whether to react to a port create event which is not available when
-        # this method is called.
-
-        return
 
     def _get_physnet_from_link_info(self, port, physnet_info):
 
@@ -464,7 +445,7 @@ class AristaDriver(driver_api.MechanismDriver):
                 new_status == n_const.PORT_STATUS_DOWN):
             LOG.debug("Handling port migration for: %s " % orig_port)
             network_id = orig_port['network_id']
-            tenant_id = orig_port['tenant_id'] or INTERNAL_TENANT_ID
+            tenant_id = orig_port['tenant_id'] or constants.INTERNAL_TENANT_ID
             # Ensure that we use tenant Id for the network owner
             tenant_id = self._network_owner_tenant(context, network_id,
                                                    tenant_id)
@@ -504,11 +485,11 @@ class AristaDriver(driver_api.MechanismDriver):
             # 2. If segment_id is provisioned and it not bound to any port it
             # should be removed from EOS.
             network_id = orig_port['network_id']
-            tenant_id = orig_port['tenant_id'] or INTERNAL_TENANT_ID
+            tenant_id = orig_port['tenant_id'] or constants.INTERNAL_TENANT_ID
             # Ensure that we use tenant Id for the network owner
             tenant_id = self._network_owner_tenant(context, network_id,
                                                    tenant_id)
-            for binding_level in context._original_binding_levels:
+            for binding_level in context._original_binding_levels or []:
                 if self._network_provisioned(
                         context, tenant_id, network_id,
                         segment_id=binding_level.segment_id):
@@ -538,7 +519,7 @@ class AristaDriver(driver_api.MechanismDriver):
                                 plugin_context, tenant_id, network_id,
                                 binding_level.segment_id)
                         except arista_exc.AristaRpcError:
-                            LOG.info(EOS_UNREACHABLE_MSG)
+                            LOG.info(constants.EOS_UNREACHABLE_MSG)
 
             return True
 
@@ -577,7 +558,7 @@ class AristaDriver(driver_api.MechanismDriver):
         # device_id and device_owner are set on VM boot
         port_id = new_port['id']
         network_id = new_port['network_id']
-        tenant_id = new_port['tenant_id'] or INTERNAL_TENANT_ID
+        tenant_id = new_port['tenant_id'] or constants.INTERNAL_TENANT_ID
         # Ensure that we use tenant Id for the network owner
         tenant_id = self._network_owner_tenant(context, network_id, tenant_id)
 
@@ -671,7 +652,7 @@ class AristaDriver(driver_api.MechanismDriver):
         port_id = port['id']
         port_name = port['name']
         network_id = port['network_id']
-        tenant_id = port['tenant_id'] or INTERNAL_TENANT_ID
+        tenant_id = port['tenant_id'] or constants.INTERNAL_TENANT_ID
         # Ensure that we use tenant Id for the network owner
         tenant_id = self._network_owner_tenant(context, network_id, tenant_id)
         sg = port['security_groups']
@@ -747,7 +728,7 @@ class AristaDriver(driver_api.MechanismDriver):
                 except ml2_exc.MechanismDriverError:
                     # If deleting a port fails, then not much can be done
                     # about it. Log a warning and move on.
-                    LOG.warning(UNABLE_TO_DELETE_PORT_MSG)
+                    LOG.warning(constants.UNABLE_TO_DELETE_PORT_MSG)
             if port_provisioned and net_provisioned and hostname and \
                 is_vm_boot and not port_down:
                 LOG.info(_LI("Port plugged into network"))
@@ -807,7 +788,7 @@ class AristaDriver(driver_api.MechanismDriver):
         host = context.host
         network_id = port['network_id']
 
-        tenant_id = port['tenant_id'] or INTERNAL_TENANT_ID
+        tenant_id = port['tenant_id'] or constants.INTERNAL_TENANT_ID
         # Ensure that we use tenant Id for the network owner
         tenant_id = self._network_owner_tenant(context, network_id, tenant_id)
 
@@ -825,7 +806,7 @@ class AristaDriver(driver_api.MechanismDriver):
         except ml2_exc.MechanismDriverError:
             # Can't do much if deleting a port failed.
             # Log a warning and continue.
-            LOG.warning(UNABLE_TO_DELETE_PORT_MSG)
+            LOG.warning(constants.UNABLE_TO_DELETE_PORT_MSG)
 
     def _delete_port(self, context, port, host, tenant_id, segments=None):
         """Deletes the port from EOS.
@@ -848,7 +829,7 @@ class AristaDriver(driver_api.MechanismDriver):
         sg = port['security_groups']
 
         if not device_id or not host:
-            LOG.warning(UNABLE_TO_DELETE_DEVICE_MSG)
+            LOG.warning(constants.UNABLE_TO_DELETE_DEVICE_MSG)
             return
 
         try:
@@ -866,7 +847,7 @@ class AristaDriver(driver_api.MechanismDriver):
             # if necessary, delete tenant as well.
             self.delete_tenant(context, tenant_id)
         except arista_exc.AristaRpcError:
-            LOG.info(EOS_UNREACHABLE_MSG)
+            LOG.info(constants.EOS_UNREACHABLE_MSG)
 
     def _delete_segment(self, context, tenant_id):
         """Deletes a dynamic network segment from EOS.
@@ -907,7 +888,7 @@ class AristaDriver(driver_api.MechanismDriver):
                                                       tenant_id, network_id,
                                                       binding_level.segment_id)
                     except arista_exc.AristaRpcError:
-                        LOG.info(EOS_UNREACHABLE_MSG)
+                        LOG.info(constants.EOS_UNREACHABLE_MSG)
                 else:
                     LOG.debug("Cannot delete segment_id %(segid)s "
                               "segment is %(seg)s",
@@ -920,10 +901,7 @@ class AristaDriver(driver_api.MechanismDriver):
         If this port is the last port using the segmentation id allocated
         by the driver, it should be released
         """
-        if migration:
-            host = context.original_host
-        else:
-            host = context.host
+        host = context.original_host if migration else context.host
 
         physnet_info = self.eapi.get_physical_network(host, context=context)
         physnet = physnet_info.get('physnet')
@@ -955,7 +933,8 @@ class AristaDriver(driver_api.MechanismDriver):
         # When Arista driver participate in port binding by allocating dynamic
         # segment and then calling continue_binding, the driver should the
         # second last driver in the bound drivers list.
-        if segment_id and bound_drivers[-2:-1] == [MECHANISM_DRV_NAME]:
+        if (segment_id and bound_drivers[-2:-1] ==
+                [constants.MECHANISM_DRV_NAME]):
             filters = {'segment_id': segment_id}
             result = db_lib.get_port_binding_level(plugin_context, filters)
             LOG.debug("Looking for entry with filters=%(filters)s "
@@ -985,7 +964,7 @@ class AristaDriver(driver_api.MechanismDriver):
             try:
                 self.rpc.delete_tenant(tenant_id)
             except arista_exc.AristaRpcError:
-                LOG.info(EOS_UNREACHABLE_MSG)
+                LOG.info(constants.EOS_UNREACHABLE_MSG)
                 raise ml2_exc.MechanismDriverError(method='delete_tenant')
 
     def _host_name(self, hostname):
@@ -1121,12 +1100,12 @@ def cli():
     api_type = confg['api_type'].upper()
     if api_type == 'EAPI':
         LOG.info("Using EAPI for RPC")
-        rpc = arista_ml2.AristaRPCWrapperEapi(ndb)
+        rpc = ml2_rpc.AristaRPCWrapperEapi(ndb)
     elif api_type == 'JSON':
         LOG.info("Using JSON for RPC")
-        rpc = arista_ml2.AristaRPCWrapperJSON(ndb)
+        rpc = ml2_rpc.arista_json.AristaRPCWrapperJSON(ndb)
     elif api_type == 'NOCVX':
-        rpc = arista_ml2.AristaNoCvxWrapperBase(ndb)
+        rpc = ml2_rpc.AristaNoCvxWrapperBase(ndb)
     else:
         msg = "RPC mechanism %s not recognized" % api_type
         LOG.error(msg)
@@ -1143,7 +1122,7 @@ def cli():
             join(Port.port_binding). \
             join(Port.port_binding_levels). \
             options(joinedload(Port.security_groups)). \
-            filter(PortBindingLevel.driver == MECHANISM_DRV_NAME). \
+            filter(PortBindingLevel.driver == constants.MECHANISM_DRV_NAME). \
             options(contains_eager(Port.port_binding_levels))
 
         if config.CONF.port_id:
